Error responses now provide "code"
Hrana Error object supports an optional "code", which is usually the
SQLite error name.

Since we're using the C library directly, we can't get the name, so
get the code and convert to the name using a dict.

To keep in sync, use the C header and parse it using 2 reg-exp.

ResponseError can also understand SqliteError as input parameter as
well as providing a tojson(), making it easier to use.